Refactor and streamline analysis scripts. Adding correlation analysis to Study plot. Close #125
