+++ Ressources/ClassroomConfigs/ClassroomConfig.json
@@ -1,12 +1,16 @@
 {
-    "name":"None-ADHD",
+    "name":"Test",
     "seed":42,
     "ticks":1000,
     "timescale": 1.0,
     "agent_types":
     [
+        {"name":"Ambitious", "openess":0.6,"conscientousness":0.8,"extraversion":0.8,"agreeableness":0.2,"neuroticism":0.2},
+        {"name":"Difficult", "openess":0.6,"conscientousness":0.8,"extraversion":0.8,"agreeableness":0.2,"neuroticism":0.8},
         {"name":"ADHD", "openess":-1.0,"conscientousness":0.2,"extraversion":-1.0,"agreeableness":0.2,"neuroticism":0.8},
-        {"name":"Normal", "openess":0.75,"conscientousness":0.6,"extraversion":0.55,"agreeableness":0.65,"neuroticism":0.50}
+        {"name":"Engineer", "openess":0.7,"conscientousness":0.8,"extraversion":0.2,"agreeableness":0.5,"neuroticism":0.3},
+        {"name":"Peacekeeper", "openess":-1.0,"conscientousness":0.6,"extraversion":0.7,"agreeableness":0.9,"neuroticism":0.2},
+        {"name":"Normal", "openess":0.6,"conscientousness":0.5,"extraversion":0.6,"agreeableness":0.5,"neuroticism":0.5}
     ],
-    "nAgents":[0,30]
+    "nAgents":[5, 5, 5, 5, 5, 5]
 }
