+++ analysis/generatePlots.py
@@ -37,21 +37,7 @@ agentBehaviorsLabels = ['B(I)', 'B(T)', 'B(W)', 'B(E)',
 NUM_ACTION_STATES = 4
 behavior_colors = ['#512bc4', '#c4892b', '#9ec42b', '#c42b9e', '#5dc7b7']
 
-
-def main(argv):
-    try:
-        classroom_stats_file = argv[1]
-        agents_stats_file = argv[2]
-        project_name = argv[3]
-    except:
-        print('%s [ClassroomStats.csv] [AgentStats.csv] [OutputPath]' % argv[0])
-        sys.exit(1)
-
-    generatePlots(classroom_stats_file, agents_stats_file, project_name)
-    print('Finished!')
-
-
-def generatePlots(classroom_stats_file, agents_stats_file, output_folder, skip_agent_plots=False):
+def generatePlots(classroom_stats, agents_stats, agent_infos, output_folder, skip_agent_plots=False):
     """Generate all plots of a single experiment (HA-Plot, Classroom Aggregats and AgentInfo)
     
     Arguments:
@@ -62,16 +48,12 @@ def generatePlots(classroom_stats_file, agents_stats_file, output_folder, skip_a
     Keyword Arguments:
         skip_agent_plots {bool} -- [description] (default: {False})
     """
-    classroom_stats = pd.read_csv(classroom_stats_file)
-    agents_stats = pd.read_csv(agents_stats_file)
-
     output_folder = os.path.abspath(output_folder)
 
     agent_out = os.path.join(output_folder, 'HA-Plot.png')
     print('Plot Happiness vs Attention Plot to [%s] ... ' % agent_out)
 
-    # Calculate Agent Info
-    agent_infos, agents_stats = calculate_agent_info(agents_stats)
+    # Calculate plot limits based on means
     max_mean_durations = agent_infos['mean_durations'].apply(max).max() * 1.1
     max_relative_durations = agent_infos['relative_durations'].apply(max).max() * 1.1
 
@@ -91,82 +73,10 @@ def generatePlots(classroom_stats_file, agents_stats_file, output_folder, skip_a
             print('Generating Agent Info plot: %s ...' % agent_info_out)
             plotAgentInfo(agent, info, agent_info_out, agentBehaviorsLabels, behavior_colors, mean_ylimits=(0.0, max_mean_durations), relative_ylimits=(0.0, max_relative_durations))
 
-    # Add the number of studying students to the classroom stats
-    behavior_sums = agents_stats[['Turn', 'IsStudying', 'IsQuarrel']].groupby('Turn').sum().astype(int)
-    classroom_stats = classroom_stats.join(behavior_sums, on='Turn').rename(columns={'IsStudying': 'Studying_sum', 'IsQuarrel': 'Quarrel_sum'})
-
     classroom_out = os.path.join(output_folder, 'ClassroomAggregates.png')
     print('Plot Classroom Aggregates to [%s] ...' % classroom_out)
     plotClassroomAggregates(classroom_stats, classroom_out)
-
-    # Store summary data into csv
-    write_experiment_summary(classroom_stats, agents_stats, output_folder)
-
-
-def calculate_agent_info(agents_stats):
-    print('Calculating Agent infos ...')
-    # First line in agent stats are personality traits
-    #meta_df = agents_stats[agents_stats['Turn'] == -2][['Tag', 'Motivation', 'Happiness', 'Attention', 'Action', 'Desire']].reset_index(drop=True)
-    meta_df = agents_stats[agents_stats['Turn'] == -2][['Tag', 'Motivation', 'Happiness']].reset_index(drop=True)
-    meta_df.rename(columns={'Motivation': 'studentname', 'Happiness': 'personalitytype', 'Attention': 'x', 'Action': 'x', 'Desire': 'x'}, inplace=True)
-    meta_df.set_index('Tag', inplace=True)
-
-    personalities_df = agents_stats[agents_stats['Turn'] == -1][['Tag', 'Motivation', 'Happiness', 'Attention', 'Action', 'Desire']].reset_index(drop=True)
-    personalities_df.rename(columns={'Motivation': 'Openess', 'Happiness': 'Conscientiousness', 'Attention': 'Extraversion', 'Action': 'Agreeableness', 'Desire': 'Neuroticism'}, inplace=True)
-    personalities_df.set_index('Tag', inplace=True)
-    personalities = personalities_df.apply(lambda x: ', '.join(['%s: %s'%(k, v) for k, v in x.to_dict().items()]), axis=1)
-    pdf = pd.DataFrame(index=personalities.index, data=personalities.values, columns=['personality'])
-
-    pdf = pd.concat([meta_df, pdf], axis=1)
-
-    # Remove all lines that are no stats, and set their datatype
-    agents_stats = agents_stats[agents_stats['Turn'] >= 0].astype({'Action': 'str', 'Desire': 'str', 'Tag':'str', 'Motivation': 'float', 'Attention': 'float', 'Happiness': 'float'})
-
-    # Get Action indices
-    agents_stats['Action_idx'] = agents_stats['Action'].apply(lambda x: identifyAction(x, agentBehaviors))
-    agents_stats['Desire_idx'] = agents_stats['Desire'].apply(lambda x: identifyAction(x, agentBehaviors))
-    agent_infos = agents_stats.groupby('Tag').apply(getAgentInfos)
-
-    # Add one field indicating if agent is studying
-    agents_stats['IsStudying'] = agents_stats['Action_idx'].isin(range(8, 8+8))
-    agents_stats['IsQuarrel'] = agents_stats['Action_idx'].isin(range(16, 16+4)) 
-
-    agent_infos = pd.concat([agent_infos, pdf], axis=1)
-
-    return agent_infos, agents_stats
-
-def write_experiment_summary(classroom_stats, agents_stats, output_folder):
-    summary_file = os.path.join(os.path.dirname(output_folder), 'Experiment_summary.csv')
-    print('Writing Experiment summary file to %s ...' % summary_file)
-
-    classroom_means = classroom_stats[['Tag', 'Motivation_mean', 'Happiness_mean', 'Attention_mean']].rename({'Motivation_mean':'Motivation', 'Happiness_mean':'Happiness', 'Attention_mean':'Attention'}, axis=1)
-    classroom_means = classroom_means.groupby('Tag').mean()
-
-   # Attention is calculated only during studying
-    agent_attention = agents_stats[agents_stats['IsStudying']][['Tag', 'Attention']].groupby('Tag').mean()
-    agent_means = agents_stats[agents_stats['Turn'] > 0][['Tag', 'Happiness', 'Motivation' ]].groupby('Tag').mean()
-    agent_means = pd.concat([agent_attention, agent_means], axis=1)
-
-    #agent_means = agents_stats[agents_stats['Turn'] > 0][['Tag', 'Motivation', 'Happiness', 'Attention']]
-    means = pd.concat([classroom_means, agent_means], axis=0, sort=True)
-    means['Instance'] = os.path.basename(output_folder)
-    means['Experiment'] = os.path.basename(os.path.dirname(output_folder))
-    means.reset_index(inplace=True)
-    
-    if os.path.isfile(summary_file):
-        header=False
-    else:
-        header=True
-
-    with open(summary_file, 'a') as f:
-        means.to_csv(f, header=header, index=False)
-    
-
-def identifyAction(string, actions):
-    for idx, a in enumerate(actions):
-        if(string.find(a) > -1):
-            return idx
-    return -1
+   
 
 def plotHappinessAttentionGraph(attention, happiness, attention_std=None, happiness_std=None, width=None, height=None, suptitle='', labels=None, include_means=True, normalize=True, ax=None):
 
@@ -271,36 +181,6 @@ def plot_mean(x, means, ax, label, color, ylimits=(None, None)):
     plot_mean_with_std(x, means, None, ax, label, color, ylimits=ylimits)
 
 
-def getAgentInfos(table):
-    """Calculate transition matrix and action sequence durations for each agent
-    """
-    tm, sequence_durations = getTransitionInfo(table['Action_idx'].values, len(agentBehaviors))
-    relative_duration, mean_durations, mean_durations_std = calculate_duration_info(sequence_durations)
-        
-    index=['TM', 'relative_durations', 'mean_durations', 'mean_durations_std']
-    return pd.Series([tm, relative_duration, mean_durations, mean_durations_std], index)
-
-
-def calculate_duration_info(sequence_durations):
-    """Use Transition matrix to calculate serverel infos about the agent
-    """
-    # Calculate ratio of each action compared to the total
-    relative_duration = [sum(l) for l in sequence_durations]
-    relative_duration = np.array(relative_duration)
-    relative_duration = relative_duration / sum(relative_duration)
-    relative_duration *= 100.0
-
-    # Calculate mean action durations
-    mean_durations = [np.mean(d) for d in sequence_durations]
-    # Fill nan with 0.0 
-    mean_durations = np.array(mean_durations)
-    mean_durations[np.isnan(mean_durations)] = 0
-
-    # Calculate STD
-    mean_durations_std = [np.std(l) for l in sequence_durations]
-
-    return relative_duration, mean_durations, mean_durations_std
-
 
 def plotAgentInfo(name, info, output_file, actions, action_colors, mean_ylimits=(0.0, 1.0), relative_ylimits=(0.0, 1.0)):
     """
@@ -365,45 +245,3 @@ def generateActionBarsPlot(y, std, ax, actions, action_colors, ylimits=(0.0, 1.0
 
     ax.set_title(title)
     ax.set_ylabel(ylabel)
-
-
-def getTransitionInfo(sequence, nActions):
-    """Calculate the Transition Matrix and Durations of Action sequences
-    """
-    tm = np.zeros(shape=(nActions, nActions))
-
-    # Calculate the transitions
-    getTransitionMatrix(sequence, tm, rowNormalization=True)
-
-    # Calculate durations of each action sequence
-    durations = getDurations(sequence, nActions)
-
-    return tm, durations
-
-
-def getTransitionMatrix(sequence, tm, rowNormalization=True):
-    for prev, next in zip(sequence, sequence[1:]):
-        tm[prev, next] = tm[prev, next] + 1 
-    if rowNormalization:
-        ntm = tm / tm.sum(axis=1)[:, None]
-        ntm[np.isnan(ntm)] = 0
-        tm[:, :] = ntm[:, :]
-
-
-def getDurations(sequence, nActions):
-    durations = [[] for x in range(nActions)]
-
-    l = 1
-    for prev, next in zip(sequence, sequence[1:]):
-        if(prev == next):
-            l+=1
-        else:
-            durations[prev].append(l)
-            l = 1
-    # Store last subsequence length
-    durations[next].append(l)
-
-    return durations
-
-if __name__ == "__main__":
-    main(sys.argv)
